Fix nightmare op.get bug
Automerge seems to have a bug somewhere where a document can
get saved with opsets stored as a map of indexes like this:
```
ops: {
  0: <myOpset>
}
```

Instead of like this:

```
ops: [
  <myOpset>
]
```

For the life of me, I cannot get a simplified reproducable example,
though I've figured out that it's something to do the way we
were applying changes. I had copied those from DocSet, but I removed
them and it seems to work now.

Here's hoping for the best
-Evan